Added dtype option to permute_weights()

+++ general_utils/weights_utils.py
@@ -1,7 +1,7 @@
 import numpy as np
 from general_utils.user_helper_functions import get_possible_weights
 
-def permute_weights(slim_length, possible_weights = None):
+def permute_weights(slim_length, possible_weights = None, dtype = float):
     '''
     Simple function to generate permutations of weights from 0-3 for scoring a peptide motif of defined length
 
@@ -18,7 +18,7 @@ def permute_weights(slim_length, possible_weights = None):
 
     permutations_count = np.prod([len(arr) for arr in possible_weights])
 
-    result = np.empty((permutations_count, slim_length), dtype=np.int)
+    result = np.empty((permutations_count, slim_length), dtype = dtype)
     meshgrid_arrays = np.meshgrid(*possible_weights)
     for i, arr in enumerate(meshgrid_arrays):
         result[:, i] = arr.ravel()
